feat: ViewModel を実装
